Move the encryption things out of the root of security module
